% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cat_split}
\alias{cat_split}
\title{Augment cat to print to both stdout and a file.}
\usage{
cat_split(
  ...,
  file = "",
  sep = " ",
  fill = FALSE,
  labels = NULL,
  append = FALSE
)
}
\arguments{
\item{...}{Passed directly to \code{cat()}.}

\item{file}{If "", then this function behaves exactly like
\code{cat()}. If not "", then \code{...} is printed to
the file argument as well as to the stdout.}

\item{sep}{Passed directly to \code{cat()}.}

\item{fill}{Passed directly to \code{cat()}.}

\item{labels}{Passed directly to \code{cat()}.}

\item{append}{Passed directly to \code{cat()}.}
}
\description{
Wrapper around \code{base::cat()} to allow printing
to a file as well as stdout. If file is "", then
this function behaves exactly the same as a normal
\code{cat()} printing to stdout. If file is not "",
then the \code{...} arguments will be printed to the
stdout AND to the file.
}
