% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{change_plot_size}
\alias{change_plot_size}
\title{Change the default plot size.}
\usage{
change_plot_size(width = 7, height = NULL)
}
\arguments{
\item{width}{The width in inches.}

\item{height}{The height in inches. If height is NULL, a square
aspect ratio is set with \code{width} side length.}
}
\value{
Nothing.
}
\description{
Changes the default plot size. Useful for in-line display
as in Jupyter Lab. Has no relation to the size of saved
plot image files.
}
