% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_sec_to_dhms}
\alias{convert_sec_to_dhms}
\title{Convert n seconds to DD:HH:MM:SS string.}
\usage{
convert_sec_to_dhms(seconds, digits = 0, strip_zeros = FALSE)
}
\arguments{
\item{seconds}{The number of seconds to convert.}

\item{digits}{The number of digits after the decimal point to keep.}

\item{strip_zeros}{If TRUE, then leading 00: fields will be removed.}
}
\value{
The string representing the converted time.
}
\description{
Converts a numeric number of seconds to a string containing that time
in DD:HH:MM:SS format (D = days, H = hours, M = minutes, S = seconds).
}
