% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{display_palette}
\alias{display_palette}
\title{Display a color palette.}
\usage{
display_palette(palette)
}
\arguments{
\item{palette}{A vector of colors, typically as hex strings.}
}
\value{
A ggplot object showing the colors in the palette.
}
\description{
Plots a vector of colors as a stack of bars in a ggplot plot.
How good this looks is heavily dependent on the plot size and
number of colors. Intended for quick visualization, not for
reliably producing palette representations that look good.
}
