% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{localMinima}
\alias{localMinima}
\title{Return indices of local minima}
\usage{
localMinima(x, endpoints = c("neither", "start", "end", "both"))
}
\arguments{
\item{x}{The numeric vector over which minima are calculated}

\item{endpoints}{Directs if the first and/or last points should
be included as viable candidates. Options are
\code{c("neither", "start", "end", "both")}. Default is "neither"}
}
\description{
Not robust. Defines local minima as points that are
less than both their immediate neighbors. As such,
it is quite sensitive to noise.
}
