% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pretty_print_vector}
\alias{pretty_print_vector}
\title{Convert a vector to a pretty-printable string.}
\usage{
pretty_print_vector(vec, number_of_columns = 1, indent = "    ")
}
\arguments{
\item{vec}{The vector to be printed.}

\item{number_of_columns}{The number of columns to print the vector
with.}

\item{indent}{A string to be printed before every row.}
}
\value{
The pretty-printable string. Best passed to \code{cat}.
}
\description{
Converts a vector to a string that can be printed. The vector
will be printed  in columns where the column width is dynamically
chosen based on the longest element in the vector.
}
