% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_n_cores}
\alias{set_n_cores}
\title{Set the number of cores.}
\usage{
set_n_cores(number_of_cores)
}
\arguments{
\item{number_of_cores}{The number of cores to set.}
}
\value{
The number of cores successfully set.
}
\description{
Sets the number of cores option that will be used
by functions that check it. i.e. \code{mc.cores}
}
